backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
requests==2.26.0
Pillow==10.1.0
djoser==2.2.2
django-filter==23.3
psycopg2-binary==2.9.3